requirements.txt
dvc
dvc[gdrive]
sklearn
tensorflow
keras
pandas
pillow